Abstract scrolling list for tracks and activities
